requirements.txt
aioboto3==13.0.1
aiobotocore==2.13.0
aiofiles==23.2.1
aiohttp==3.9.4
aioitertools==0.11.0
aiosignal==1.3.1
altair==5.3.0
annotated-types==0.7.0
anthropic==0.28.1
anyio==4.4.0
appdirs==1.4.4
async-timeout==4.0.3
asyncio==3.4.3
attrs==23.2.0
beautifulsoup4==4.12.3
blinker==1.7.0
boto3==1.34.106
botocore==1.34.106
cachetools==5.3.3
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
contourpy==1.2.1
cryptography==42.0.5
cycler==0.12.1
dataclasses-json==0.6.7
datasets==2.19.2
dill==0.3.8
distro==1.9.0
entrypoints==0.4
exceptiongroup==1.2.1
Faker==24.7.1
favicon==0.7.0
filelock==3.15.1
fonttools==4.51.0
frozenlist==1.4.1
fsspec==2024.3.1
gitdb==4.0.11
GitPython==3.1.43
h11==0.14.0
htbuilder==0.6.2
httpcore==1.0.5
httpx==0.27.0
huggingface-hub==0.23.4
idna==3.7
Jinja2==3.1.3
jmespath==1.0.1
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
kiwisolver==1.4.5
langchain==0.2.2
langchain-aws==0.1.5
langchain-community==0.2.4
langchain-core==0.2.9
langchain-openai==0.1.8
langchain-text-splitters==0.2.1
langsmith==0.1.79
lxml==5.2.1
Markdown==3.6
markdown-it-py==3.0.0
markdownlit==0.0.7
MarkupSafe==2.1.5
marshmallow==3.21.3
matplotlib==3.8.4
mdurl==0.1.2
more-itertools==10.2.0
multidict==6.0.5
multiprocess==0.70.16
mypy-extensions==1.0.0
nest-asyncio==1.6.0
numpy==1.26.4
openai==1.34.0
orjson==3.10.5
packaging==24.0
palettable==3.3.3
pandas==2.2.1
pillow==10.3.0
pip==23.3
prometheus_client==0.20.0
protobuf==4.25.3
pyarrow==15.0.2
pyarrow-hotfix==0.6
pycparser==2.22
pydantic==2.7.4
pydantic_core==2.18.4
pydeck==0.8.1b0
Pygments==2.17.2
pymdown-extensions==10.7.1
pyparsing==3.1.2
pypdf==4.2.0
pysbd==0.3.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
ragas==0.1.9
referencing==0.34.0
regex==2024.5.15
requests==2.32.3
rich==13.7.1
rpds-py==0.18.0
s3transfer==0.10.1
setuptools==68.2.0
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
soupsieve==2.5
SQLAlchemy==2.0.30
st-annotated-text==4.0.1
streamlit==1.33.0
streamlit-camera-input-live==0.2.0
streamlit-card==1.0.0
streamlit-embedcode==0.1.2
streamlit-extras==0.4.2
streamlit-faker==0.0.3
streamlit-image-coordinates==0.1.6
streamlit-keyup==0.2.4
streamlit-toggle-switch==1.0.2
streamlit-vertical-slider==2.5.5
tenacity==8.2.3
tiktoken==0.7.0
toml==0.10.2
toolz==0.12.1
tornado==6.4
tqdm==4.66.4
typing-inspect==0.9.0
typing_extensions==4.11.0
tzdata==2024.1
urllib3==2.0.7
validators==0.28.0
wheel==0.41.2
wrapt==1.16.0
xxhash==3.4.1
yarl==1.9.4